# api/openapi/openapi.yaml
openapi: 3.1.1
info:
  title: sshpilot-vallete API
  version: 1.0.0-rc1
  summary: Publishes SSH public key sets as native authorized_keys documents.
  description: |
    The contract for what the server implements **today**: the two probe
    endpoints, the unauthenticated publish read path, and device management.

    The remaining management endpoints (owner onboarding, handle and key-set
    administration, device pairing, credential enrollment) are **forthcoming**
    and are deliberately absent. A specification that describes unbuilt routes
    cannot be used as a contract, because a consumer has no way to tell the
    parts that are real from the parts that are aspirational. Routes are added
    here as they land, and a contract test fails the build when this document
    and the router disagree in either direction.

    ## Errors on the management API

    Every management error response is the same JSON object and carries no
    diagnostics:

    ```json
    { "status": "error" }
    ```

    This is documented as it actually is. The body does not name a field, a
    reason, or a code, and no error-detail model is described here, because the
    server returns none — a richer schema in this document would promise
    consumers a level of detail they will never receive.

    ## Authorization

    Management endpoints require a bearer token. The owner is taken from that
    token and from nowhere else: **no management route accepts an owner
...

...

...

...

    ## Transport

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/v1/admin/owners:
    post:
      tags: [admin]
      operationId: provisionOwner
      summary: Provision a new owner and return a one-time enrollment code.
      description: |
        Creates an owner, claims its handle, and opens an empty public default
        key set — in one transaction — on an administrator's authority, then
        returns a short-lived, single-use enrollment code (ADR-0033, Phase-1
        decision #14). The owner redeems that code at `POST /api/v1/enroll/redeem`
        to mint their own refresh and access tokens, so the operator never holds
        the owner's long-lived credential.

        The request is authorized against an ACTIVE administrator, audited, and
        persisted before the owner exists; a disabled administrator's token is
        refused even though its signature verifies (ADR-0031). The handle and any
        explicit set name pass the reserved-identifier blocklist (ADR-0017); a
        refused name and a syntactically invalid one both answer `400`, rendered
        the same so nothing reveals which rule fired. A `409` means the handle is
        taken. The one-time code is returned in the response body only and is
        never written to the audit log.
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionOwnerRequest"
      responses:
        "201":
          $ref: "#/components/responses/OwnerProvisioned"
        "400":
          $ref: "#/components/responses/AdminEditBadRequest"
        "403":
          $ref: "#/components/responses/AdminRefused"
        "409":
          $ref: "#/components/responses/OwnerHandleConflict"
        "429":
          $ref: "#/components/responses/AdminRateLimited"
        "500":
          $ref: "#/components/responses/ManagementInternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        An access token issued by the enrollment flow. Exactly one
        `Authorization` header is accepted; two is a request-smuggling shape and
        is refused rather than resolved.
    adminAuth:
      type: http
      scheme: bearer
      description: |
        A signed administrator bearer token (ADR-0031). The token names an
        administrator and is signed with a DEDICATED administrator signing key,
        distinct from the owner access-token signing key, so an owner token can
        never grant administrator authority (ADR-0018) and an owner-key
        compromise cannot forge one. Provision the first administrator and mint
        its token with the `bootstrap-admin` subcommand.

        Verification checks the signature and validity window only; the actual
        authority is the `administrators` table, consulted on every request, so
        a validly-signed token for a disabled or unknown administrator is
        refused with `403`. There is no per-token revocation in v1: disabling
        the administrator row revokes all of that admin's tokens at once, and a
        bounded token lifetime caps a leaked token's exposure until then.

        Whether the server accepts these tokens depends on deployment wiring: an
        administrator signing key must be configured
        (`auth.admin_token_signing_key_ref`), required in production. Without it
        the admin routes stay fail-closed and every edit is refused with `403`.
...
  headers:
...
    CacheControlNoStore:
      description: |
        `no-store`. Failure and probe responses are never cached. A cached
        negative would keep answering `404` after the owner published the set,
        and a cached probe would let an unhealthy instance keep passing.
      required: true
      schema:
        type: string
...
    VaryAuthorization:
      description: |
        Always `Authorization`. It is set on every access-gated response,
        success and refusal alike (ADR-0019). The success path is where it
        matters most: two owners issuing the same request each see a body they
        alone may see, and a shared cache without this header would serve the
        first one's devices to the second.
      required: true
      schema:
        type: string
...
  responses:
...
    ManagementInternalError:
      description: |
        An internal fault, including a management route reached on a deployment
        where the service behind it was not wired. The cause is in the server
        log and never on the wire.

        A wiring fault is deliberately a `500` and not a `404`: degrading into
        "not found" would be indistinguishable from a genuine miss and would
        hide a broken deployment behind a plausible answer.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Vary:
          $ref: "#/components/headers/VaryAuthorization"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { status: error }
...
    AdminRefused:
      description: |
        The edit was refused because the request did not establish an active
        administrator — no administrator identity, an unknown one, or a disabled
        one. These cases are deliberately indistinguishable: telling them apart
        would let a caller enumerate which administrator IDs exist. The body is
        fixed text and names nothing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { status: error }
    AdminEditBadRequest:
      description: |
        The request body is unreadable, oversized, carries an unrecognized
        field, or names an entry that is empty or has no comparable skeleton. It
        reports only on what the caller sent, and names no field and no reason.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { status: error }
...
    OwnerProvisioned:
      description: |
        The owner, its handle, and its default key set were created, and a
        one-time enrollment code was minted. The code is disclosed HERE and
        nowhere else — it is a credential and never appears in the audit log.
        The administrator hands it to the owner, who redeems it once at
        `POST /api/v1/enroll/redeem` before it expires.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProvisionedOwner"
    OwnerHandleConflict:
      description: |
        The requested handle is already claimed, by this owner or another. The
        answer is the same either way and names no owner: a handle that resolves
        to another owner must be indistinguishable from one that is simply taken.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { status: error }
    AdminRateLimited:
      description: |
        The ADMIN rate-limit tier refused the request: the administrator's key
        has exhausted its budget for the window, or the counter store was
        unavailable and the tier fails closed (ADR-0023). The body names no
        count, limit, or tier, for the same reason the other tiers' refusals do
        not.
      headers:
        Retry-After:
          description: |
            Seconds until the administrator's window resets, rounded up. Derived
            from the actual remaining life of the window, not a constant.
          required: true
          schema:
            type: integer
            minimum: 0
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { status: error }
...
  schemas:
...
    Error:
      type: object
      description: |
        The uniform management error body. It carries a single constant field
        and no diagnostics.

        There is deliberately no code, message, or field pointer. Every error on
        this surface is this exact object, which is what makes a `404` for
        another owner's device byte-identical to a `404` for a device that never
        existed.
      required: [status]
      properties:
        status:
          type: string
          enum: [error]
      additionalProperties: false
...
    ProvisionOwnerRequest:
      type: object
      description: |
        The owner to provision. There is deliberately no owner id, no status, no
        key, and no device field: the server mints the identifiers, the owner is
        always created active, and the owner adds their own keys after they
        enroll. `additionalProperties: false` refuses any other field rather than
        ignoring it.
      required: [handle]
      properties:
        handle:
          type: string
          description: |
            The public name to claim for the new owner. It passes the
            reserved-identifier blocklist and the handle syntax rules; a refused
            or invalid handle is a `400` that names no reason.
        set_name:
          type: string
          description: |
            The name of the owner's default key set. Optional; omitted means
            `default`. When supplied it is checked against the blocklist too.
        client_label:
          type: string
          description: |
            A non-secret label recorded on the minted enrollment credential and
            its audit trail. Optional; omitted means a server default.
      additionalProperties: false
    ProvisionedOwner:
      type: object
      description: |
        The provisioned owner and its one-time enrollment code. The code is the
        only secret here and appears in no other response and no log.
      required: [owner_id, handle, set_name, enrollment_code, expires_at, pairing_id]
      properties:
        owner_id:
          type: string
          description: The identifier of the newly created owner.
        handle:
          type: string
          description: The handle claimed for the owner.
        set_name:
          type: string
          description: The name the default key set was created with.
        enrollment_code:
          type: string
          description: |
            The single-use code the owner redeems at `POST /api/v1/enroll/redeem`
            to mint their own tokens. It expires at `expires_at`.
        expires_at:
          type: string
          format: date-time
          description: When the enrollment code expires if unredeemed.
        pairing_id:
          type: string
          description: |
            Identifies the pairing the code belongs to. Not a secret; on its own
            it authenticates nothing.
      additionalProperties: false
...